Refactor trace_utils.rs into smaller units.
