+++ trace-mini-agent/src/trace_processor.rs
@@ -213,7 +213,7 @@ mod tests {
 
         assert_eq!(
             expected_tracer_payload,
-            tracer_payload.unwrap().get_payloads()[0]
+            tracer_payload.unwrap().tracer_payloads[0]
         );
     }
 
@@ -283,7 +283,7 @@ mod tests {
         };
         assert_eq!(
             expected_tracer_payload,
-            tracer_payload.unwrap().get_payloads()[0]
+            tracer_payload.unwrap().tracer_payloads[0]
         );
     }
 }
