+++ trace-utils/src/lib.rs
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod config_utils;
+pub mod send_data;
 pub mod stats_utils;
 pub mod trace_test_utils;
 pub mod trace_utils;
+pub mod tracer_header_tags;
